Fix dependency

+++ test/JDBC/input/babel_test_int8_numeric_oper_before_16_3-vu-prepare.sql
@@ -1,5 +1,5 @@
 -- tsql
-create table test_int8_numeric_vu(a int);
+create table test_int8_numeric_vu(a bigint);
 GO
 
 -- insert 1M rows of data
